+++ cloud/cfg/template.go
@@ -51,6 +51,13 @@ initialize_docker() {
 }
 
 initialize_minion() {
+	# Create the TLS directory now so that it will exist when the minion starts,
+	# and attempts to mount it as a volume. If the directory didn't exist, then
+	# Docker would automatically create it, resulting in it being owned by root.
+	# The TLS credential installation code running on the daemon would then be
+	# unable to write to the directory.
+	install -d -o kelda -m 755 {{.TLSDir}}
+
 	cat <<- EOF > /etc/systemd/system/minion.service
 	[Unit]
 	Description=Kelda Minion
@@ -66,7 +73,8 @@ initialize_minion() {
 	-v /var/run/docker.sock:/var/run/docker.sock \
 	-v /etc/ssl/certs/ca-certificates.crt:/etc/ssl/certs/ca-certificates.crt \
 	-v /home/kelda/.ssh:/home/kelda/.ssh:rw \
-	-v /run/docker:/run/docker:rw {{.DockerOpts}} {{.KeldaImage}} \
+	-v {{.TLSDir}}:{{.TLSDir}}:ro \
+	-v /run/docker:/run/docker:rw {{.KeldaImage}} \
 	kelda -l {{.LogLevel}} minion {{.MinionOpts}}
 	Restart=on-failure